+++ deployment/common/changeset/evm/mcms/mcms.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/smartcontractkit/chainlink/deployment"
 
-	"github.com/smartcontractkit/chainlink/deployment/common/changeset/internal/ops"
-	"github.com/smartcontractkit/chainlink/deployment/common/changeset/internal/seqs"
+	"github.com/smartcontractkit/chainlink/deployment/common/changeset/evm/mcms/ops"
+	"github.com/smartcontractkit/chainlink/deployment/common/changeset/evm/mcms/seqs"
 	"github.com/smartcontractkit/chainlink/deployment/common/changeset/state"
 	"github.com/smartcontractkit/chainlink/deployment/common/opsutils"
 	"github.com/smartcontractkit/chainlink/deployment/common/proposalutils"
